Simulate incoming message/edit for incoming reactions
Allow retracting older reaction messages, do not allow editing any of them

+++ src/main/java/eu/siacs/conversations/parser/MessageParser.java
@@ -441,7 +441,7 @@ public class MessageParser extends AbstractParser implements OnMessagePacketRece
                 }
             }
         }
-        final LocalizedContent body = packet.getBody();
+        LocalizedContent body = packet.getBody();
 
         final Element axolotlEncrypted = packet.findChildEnsureSingle(XmppAxolotlMessage.CONTAINERTAG, AxolotlService.PEP_PREFIX);
         int status;
@@ -497,6 +497,28 @@ public class MessageParser extends AbstractParser implements OnMessagePacketRece
             } else {
                 invite.execute(account);
                 return;
+            }
+        }
+
+        final Element reactions = packet.findChild("reactions", "urn:xmpp:reactions:0");
+        if (body == null && html == null) {
+            if (reactions != null && reactions.getAttribute("id") != null) {
+                final Conversation conversation = mXmppConnectionService.find(account, counterpart.asBareJid());
+                if (conversation != null) {
+                    final Message reactionTo = conversation.findMessageWithRemoteIdAndCounterpart(reactions.getAttribute("id"), null);
+                    if (reactionTo != null) {
+                        String bodyS = reactionTo.reply().getBody();
+                        for (Element el : reactions.getChildren()) {
+                            if (el.getName().equals("reaction") && el.getNamespace().equals("urn:xmpp:reactions:0")) {
+                                bodyS += el.getContent();
+                            }
+                        }
+                        body = new LocalizedContent(bodyS, "en", 1);
+                        final Message previousReaction = conversation.findMessageReactingTo(reactions.getAttribute("id"), counterpart);
+Log.d("WUT", "" + previousReaction + "    " + counterpart);
+                        if (previousReaction != null) replacementId = previousReaction.replyId();
+                    }
+                }
             }
         }
 
@@ -637,6 +659,7 @@ public class MessageParser extends AbstractParser implements OnMessagePacketRece
                 }
             }
             message.markable = packet.hasChild("markable", "urn:xmpp:chat-markers:0");
+            if (reactions != null) message.addPayload(reactions);
             for (Element el : packet.getChildren()) {
                 if ((el.getName().equals("query") && el.getNamespace().equals("http://jabber.org/protocol/disco#items") && el.getAttribute("node").equals("http://jabber.org/protocol/commands")) ||
                     (el.getName().equals("fallback") && el.getNamespace().equals("urn:xmpp:fallback:0"))) {
@@ -693,7 +716,7 @@ public class MessageParser extends AbstractParser implements OnMessagePacketRece
                             replacedMessage.setBody(message.getBody());
                             replacedMessage.putEdited(replacedMessage.getRemoteMsgId(), replacedMessage.getServerMsgId());
                             replacedMessage.setRemoteMsgId(remoteMsgId);
-                            if (!replaceElement.getName().equals("replace")) {
+                            if (replaceElement != null && !replaceElement.getName().equals("replace")) {
                                 mXmppConnectionService.getFileBackend().deleteFile(replacedMessage);
                                 mXmppConnectionService.evictPreview(message.getUuid());
                                 replacedMessage.clearPayloads();
